feat: add checks to add edit and delete product views and hide options on templates so only superusers can see or use them

+++ products/views.py
@@ -71,6 +71,9 @@ def add_product(request):
     """
     Add a product to the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to do that.')
+        return redirect('home')
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -97,6 +100,9 @@ def edit_product(request, product_id):
     """
     A view to edit/update product information
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to do that.')
+        return redirect('home')
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -125,6 +131,9 @@ def delete_product(request, product_id):
     """
     A view to delete a selected product
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to do that.')
+        return redirect('home')
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product successfully deleted.')
